requirements.txt
asgiref==3.9.1
Django==5.2.4
django-environ==0.12.0
djangorestframework==3.16.1
pillow==11.3.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
